% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_full_join.R
\name{animate_full_join}
\alias{animate_full_join}
\title{Title}
\usage{
animate_full_join(x, y, result = "gif")
}
\arguments{
\item{result}{gif or static, if static, only the last frame is shown.}
}
\description{
Title
}
\examples{
x <- data_frame(
  id = 1:3,
  x = paste0("x", 1:3)
)

y <- data_frame(
  id = (1:4)[-3],
  y = paste0("y", (1:4)[-3])
)

# render the static image
animate_full_join(x, y, "static")

\donttest{
 # render a gif
 animate_full_join(x, y)

 # to save the gif you can use
 fj <- animate_full_join(x, y)
 anim_save(here::here("images", "full-join.gif"), fj)
}
}
