% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_intersect.R
\name{animate_intersect}
\alias{animate_intersect}
\title{Animates an intersect between two datasets}
\usage{
animate_intersect(x, y, result = "gif")
}
\arguments{
\item{x}{the left dataset to intersect}

\item{y}{the right dataset to intersect}

\item{result}{gif or static, if static, only the last frame is shown.}
}
\value{
a gif or a ggplot image
}
\description{
Animates an intersect between two datasets
}
\examples{
x <- data_frame(
  id = c(1, 2, 3),
  x = c(1, 1, 2),
  y = c("a", "b", "a")
)

y <- data_frame(
  id = c(1, 4),
  x = c(1, 2),
  y = c("a", "b")
)

# render the static image
animate_intersect(x, y, "static")

\donttest{
to save a png of the static image, use
ins <- animate_intersect(x, y, "static")
ggsave("intersect.png", ins)

 # render a gif
 animate_intersect(x, y)

 # to save the gif, use
 ins <- animate_intersect(x, y)
 anim_save("intersect.png.gif", ins)
}
}
