% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_left_join.R
\name{animate_left_join}
\alias{animate_left_join}
\title{Animates a left join between two datasets}
\usage{
animate_left_join(x, y, result = "gif", title = "left_join(x, y)", ...)
}
\arguments{
\item{x}{the left dataset to join}

\item{y}{the right dataset to join}

\item{result}{gif or static, if static, only the last frame is shown.}

\item{title}{the title for the graphic}

\item{...}{Further arguments such as \code{text_family} and \code{title_family} to specify the fonts}
}
\value{
a gif or a ggplot image
}
\description{
Animates a left join between two datasets
}
\examples{
x <- data_frame(
  id = 1:3,
  x = paste0("x", 1:3)
)

y <- data_frame(
  id = (1:4)[-3],
  y = paste0("y", (1:4)[-3])
)

# render the static image
animate_left_join(x, y, "static")

\donttest{
  # to save a png of the static image, use
  lj <- animate_left_join(x, y, "static")
  ggsave("left-join.png", lj)

  # render a gif
  animate_left_join(x, y)

  # to save the gif, use
  lj <- animate_left_join(x, y)
  anim_save("left-join.gif", lj)
}
}
