% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_plot.R
\name{animate_plot}
\alias{animate_plot}
\title{Animates a Plot}
\usage{
animate_plot(x, transition_length = 2, state_length = 1)
}
\arguments{
\item{state_length}{}
}
\value{

}
\description{
Animates a Plot
}
