% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_right_join.R
\name{animate_right_join}
\alias{animate_right_join}
\title{Animates a right join between two datasets}
\usage{
animate_right_join(x, y, result = "gif", ...)
}
\arguments{
\item{x}{the left dataset to join}

\item{y}{the right dataset to join}

\item{result}{gif or static, if static, only the last frame is shown.}

\item{...}{Further arguments such as \code{text_family} and \code{title_family} to specify the fonts}
}
\value{
a gif or a ggplot image
}
\description{
Animates a right join between two datasets
}
\examples{
x <- data_frame(
  id = 1:3,
  x = paste0("x", 1:3)
)

y <- data_frame(
  id = (1:4)[-3],
  y = paste0("y", (1:4)[-3])
)

# render the static image
animate_right_join(x, y, "static")

\donttest{
  # to save a png of the static image, use
  rj <- animate_right_join(x, y, "static")
  ggsave("right-join.png", rj)

  # render a gif
  animate_right_join(x, y)

  # to save the gif, use
  rj <- animate_right_join(x, y)
  anim_save("right-join.gif", rj)
}
}
