% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_setdiff.R
\name{animate_setdiff}
\alias{animate_setdiff}
\title{Animates a setdiff operation between two datasets}
\usage{
animate_setdiff(x, y, result = "gif", ...)
}
\arguments{
\item{x}{the left dataset}

\item{y}{the right dataset}

\item{result}{gif or static, if static, only the last frame is shown.}

\item{...}{Further arguments such as \code{text_family} and \code{title_family} to specify the fonts}
}
\value{
a gif or a ggplot image
}
\description{
Animates a setdiff operation between two datasets
}
\examples{
x <- data_frame(
  id = 1:3,
  x = c(1, 1, 2),
  y = c("a", "b", "a")
)

y <- data_frame(
  id = c(1, 4),
  x = c(1, 2),
  y = c("a", "b")
)

# render the static image
animate_setdiff(x, y, "static")

\donttest{
  # to save a png of the static image, use
  setd <- animate_setdiff(x, y, "static")
  ggsave("setdiff.png", setd)

  # render a gif
  animate_setdiff(x, y)

  # to save the gif, use
  setd <- animate_setdiff(x, y)
  anim_save("setdiff.gif", setd)
}
}
