% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_union.R
\name{animate_union}
\alias{animate_union}
\title{Animates a union between two datasets}
\usage{
animate_union(x, y, result = "gif", title = "union(x, y)", ...)
}
\arguments{
\item{x}{the left dataset}

\item{y}{the right dataset}

\item{result}{gif or static, if static, only the last frame is shown.}

\item{title}{the title for the graphic}

\item{...}{Further arguments such as \code{text_family} and \code{title_family} to specify the fonts}
}
\value{
a gif or a ggplot image
}
\description{
Animates a union between two datasets
}
\examples{
x <- data_frame(
  id = 1:3,
  x = c(1, 1, 2),
  y = c("a", "b", "a")
)

y <- data_frame(
  id = c(1, 4),
  x = c(1, 2),
  y = c("a", "b")
)

# render the static image
animate_union(x, y, "static")

\donttest{
  # to save a png of the static image, use
  un <- animate_union(x, y, "static")
  ggsave("union.png", un)

  # render a gif
  animate_union(x, y)

  # to save the gif, use
  un <- animate_union(x, y)
  anim_save("union.gif", un)
}
}
