% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_union_all.R
\name{animate_union_all}
\alias{animate_union_all}
\title{Animates a union-all between two datasets}
\usage{
animate_union_all(x, y, result = "gif")
}
\arguments{
\item{x}{the left dataset}

\item{y}{the right dataset}

\item{result}{gif or static, if static, only the last frame is shown.}
}
\value{
a gif or a ggplot image
}
\description{
Animates a union-all between two datasets
}
\examples{
x <- data_frame(
  id = 1:3,
  x = c(1, 1, 2),
  y = c("a", "b", "a")
)

y <- data_frame(
  id = c(1, 4),
  x = c(1, 2),
  y = c("a", "b")
)

# render the static image
animate_union_all(x, y, "static")

\donttest{
  # to save a png of the static image, use
  un <- animate_union_all(x, y, "static")
  ggsave("union.png", un)

  # render a gif
  animate_union_all(x, y)

  # to save the gif, use
  un <- animate_union_all(x, y)
  anim_save("union.gif", un)
}
}
