% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorize_keys.R
\name{colorize_keys}
\alias{colorize_keys}
\title{Colorize Keys}
\usage{
colorize_keys(df, n_colors, key_col = "id", color_other = "#d0d0d0",
  color_missing = "#ffffff")
}
\arguments{
\item{color_missing}{}
}
\value{

}
\description{
Colorize Keys
}
