% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorize_row_id.R
\name{colorize_row_id}
\alias{colorize_row_id}
\title{Colorizes Row ID}
\usage{
colorize_row_id(df, n_colors, key_col = "id")
}
\arguments{
\item{df}{a data_frame}

\item{n_colors}{the number of colors}

\item{key_col}{the key column}
}
\value{
a data_frame
}
\description{
Colorizes Row ID
}
\examples{
NULL
}
