% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_join.R
\name{plot_data_join}
\alias{plot_data_join}
\title{Plots a join of data}
\usage{
plot_data_join(x, title = "", xlims = xlim(0.5, 5.5),
  ylims = ylim(-3.5, -0.5), ...)
}
\arguments{
\item{x}{a data_frame}

\item{title}{the title of the plot}

\item{xlims}{the xlimits}

\item{ylims}{the ylimits}

\item{...}{further argument passed to plot_data}
}
\value{
a ggplot object
}
\description{
Plots a join of data
}
\examples{
NULL
}
