% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_data.R
\name{proc_data}
\alias{proc_data}
\title{Processes the data}
\usage{
proc_data(x, .id = "x", color_fun = colorize_keys,
  color_when = c("after", "before"), ...)
}
\arguments{
\item{x}{a data_frame}

\item{.id}{the id column}

\item{color_fun}{the function to colorize the data}

\item{color_when}{when should the coloring be applied}

\item{...}{further arguments passed to color_fun}
}
\value{
a data_frame
}
\description{
Processes the data
}
\examples{
NULL
}
