% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_data_set.R
\name{proc_data_set}
\alias{proc_data_set}
\title{Processes a dataset for set operations}
\usage{
proc_data_set(x, .id = "x")
}
\arguments{
\item{.id}{}
}
\value{

}
\description{
Processes a dataset for set operations
}
