% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_static_plot.R
\name{save_static_plot}
\alias{save_static_plot}
\title{Saves a static plot}
\usage{
save_static_plot(g, filename, formats = c("png", "svg"))
}
\arguments{
\item{formats}{}
}
\value{

}
\description{
Saves a static plot
}
