% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_plot_filename.R
\name{static_plot_filename}
\alias{static_plot_filename}
\title{Constructs the filename for a file and its extension}
\usage{
static_plot_filename(x, ext)
}
\arguments{
\item{x}{a file}

\item{ext}{the extension}
}
\value{

}
\description{
Constructs the filename for a file and its extension
}
\examples{
static_plot_filename("my-file", "png")
static_plot_filename("my-file", "gif")
}
