% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_full_join.R
\name{animate_full_join}
\alias{animate_full_join}
\title{Animates a full join between two datasets}
\usage{
animate_full_join(x, y, result = "gif")
}
\arguments{
\item{x}{the left dataset to join}

\item{y}{the right dataset to join}

\item{result}{gif or static, if static, only the last frame is shown.}
}
\value{
a gif or a ggplot image
}
\description{
Animates a full join between two datasets
}
\examples{
x <- data_frame(
  id = 1:3,
  x = paste0("x", 1:3)
)

y <- data_frame(
  id = (1:4)[-3],
  y = paste0("y", (1:4)[-3])
)

# render the static image
animate_full_join(x, y, "static")

\donttest{
  # to save a png of the static image, use
  fj <- animate_full_join(x, y, "static")
  ggsave("full-join.png", fj)

  # render a gif
  animate_full_join(x, y)

  # to save the gif, use
  fj <- animate_full_join(x, y)
  anim_save("full-join.gif", fj)
}
}
