% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_join.R
\name{plot_data_join}
\alias{plot_data_join}
\title{Plots a join of data}
\usage{
plot_data_join(x, title = "", xlims = xlim(0.5, 5.5),
  ylims = ylim(-3.5, -0.5))
}
\arguments{
\item{ylims}{}
}
\description{
Plots a join of data
}
