% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_set.R
\name{plot_data_set}
\alias{plot_data_set}
\title{Plots a set operation}
\usage{
plot_data_set(x, title = "", xlims = xlim(1.5, 6.5),
  ylims = ylim(-3.5, -0.5))
}
\arguments{
\item{ylims}{}
}
\value{

}
\description{
Plots a set operation
}
