% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_data.R
\name{proc_data}
\alias{proc_data}
\title{Processes the data}
\usage{
proc_data(x, .id = "x", color_fun = colorize_keys,
  color_when = c("after", "before"), ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Processes the data
}
